configs refactor part 1/3: split _configs.py into a package
_configs.py grew to 2591 lines (9 sub-configs, 19 per-model configs,
3 mega-extractor switch functions, a 530-line ArchitectureConfig.
from_transformers). This is the first of three mechanical refactors
that carve it into a scalable package layout. No behavior change in
this PR — every public name is still importable from mobius._configs.

src/mobius/_configs/
├── __init__.py        # re-exports everything that was in _configs.py
├── _sub_configs.py    # pure-data dataclasses (RoPE/Vision/Audio/Codec/TTS)
├── _quantization.py   # QuantizationConfig + from_transformers
└── _base.py           # BaseModelConfig, ArchitectureConfig, per-model
                      # subclasses, and the _extract_* helpers

Follow-up PRs in this series:
  Part 2/3 — convert the _extract_audio_config / _extract_vision_config
             model_type switches into a decorator-registered dispatch
             so new models add a file under per_model/ instead of a
             branch in the central function.
  Part 3/3 — move per-model config subclasses (Gemma2Config,
             MllamaConfig, NemotronHConfig, ...) into per_model/ and
             carve up ArchitectureConfig.from_transformers.

Tests: 2769 passed (full src/ + tests/build_graph_test.py + cli_test.py).
Ruff: clean.
